Added Remove-VSTeamAccessControlEntry.ps1
This function removes the user and/or group from the specified resource (token) in the specified security namespace. This function handles the conversion from the AzD format of vssgp.<base64-encoded-value> to the SID based value expected by the API. Function can accept multiple AzD formatted descriptors in comma separated format for bulk-removal of users/groups from a resource.