Update VSTeamJobRequest.ps1 (#226)
* Update VSTeamJobRequest.ps1

I'm not sure if this is a bug, but I cant find anyway to get jobs that are currently running.
I get these errors because the properties does not exist on a active running build (have also verified using REST API directly):

The property 'result' cannot be found on this object. Verify that the property exists.
At C:\Program Files\WindowsPowerShell\Modules\vsTeam\6.4.1\vsteam.classes.ps1:1331 char:7
+       $this.Result = $obj.result
+       ~~~~~~~~~~~~~~~~~~~~~~~~~~
    + CategoryInfo          : NotSpecified: (:) [], PropertyNotFoundException
    + FullyQualifiedErrorId : PropertyNotFoundStrict

Same goes for "finishtime"

So in my pull-request I have included verification to check if property exists.
It's very important and critical for me to get info on how many jobs that are currently running in a pool.

Thanks for this great module! 

Best,
Asif

* Updated version, change log and help

Co-authored-by: Donovan Brown <dlb@dlbonline.com>