Pester5 (#316)
All the tests unit and integration have been ported over to Pester 5. It cut a minute off the execution time of all the unit tests. This does require the use of Pester 5 from this point forward. I am going to merge this into master but not cut a new release until a PR with new features is merged as well.