Update integration tests pipeline to use pipeline artifact
Updated Integration tests to pull down the module code from the pipeline artifact and install it. This eliminates the need to have the external nuget feed dependency. As I am unable to trigger pipeline runs for this project, I am unable to test this. 

I was able to download the module artifact and test this install method locally on my machine without issue.

Please review and let me know what you think.