fix: import-module loses previously stored configuration
This pull request resolves two main issues. First, it fixes bug #493 where the `import-module vsteam` command was causing loss of previously stored configurations. Second, it corrects the 'Versions' test failure on local setups when `Set-VSTeamAccount` is used with the `-Local` parameter. This issue didn't surface during GitHub builds due to the `TEAM_VERSION` variable not being set. In addition to these fixes, the pull request includes updates to the change log and documentation.

---------

Co-authored-by: Sebastian S Schütze <sebastian.s.schuetze@deutschebahn.com>