chore: optimized build process and updated metadata
The recent set of changes encompass optimizations to the build process and updates to the module metadata. The build pipeline was refactored into child workflows for better organization and simplicity, alongside removal of certain caching and module installation steps which were found to be less beneficial. Additionally, the version script was simplified for non-production use. On the metadata front, the psd1 file was updated with the most recent information and optimized. A few file renamings were also part of this chore to enhance clarity. These refinements collectively contribute to a more streamlined development workflow and accurate metadata representation.